increase fail penalty by 100x; resolves looping behavior
Apparently the system decided it was cheaper to not succeed than it was to try
to finish the task. Not totally sure how it came to that conclusion, but that
seemed to have been the issue. I wonder if increasing the planning depth was
another way to address this problem, since the net cost would have been greater.
Still, this is worthy of further study, as playing around with this value in
great detail isn't likely a good long-term solution.